+++ src/test/java/com/github/afanas10101111/jtb/client/GroupClientImplTest.java
@@ -13,6 +13,9 @@ import org.junit.jupiter.api.Test;
 import java.util.Collections;
 import java.util.List;
 
+import static com.github.afanas10101111.jtb.client.JavarushClientProperties.GROUPS_PATH;
+import static com.github.afanas10101111.jtb.client.JavarushClientProperties.JAVARUSH_URL;
+import static com.github.afanas10101111.jtb.client.JavarushClientProperties.POSTS_PATH;
 import static com.github.afanas10101111.jtb.client.impl.GroupClientImpl.GROUP_BY_ID_PATH;
 import static com.github.afanas10101111.jtb.client.impl.GroupClientImpl.GROUP_COUNT_PATH;
 import static com.github.afanas10101111.jtb.client.dto.GroupInfoType.TECH;
@@ -21,25 +24,21 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 class GroupClientImplTest {
-    private static final String URL = "https://javarush.ru/api/1.0/rest";
+    private static final MockClient clientMock = MockClient.register();
 
-    private final GroupClient groupClient = new GroupClientImpl(URL);
-
-    private static final String GROUPS = "/groups";
-    private static final String POSTS = "/posts";
-    private static final MockClient client = MockClient.register();
+    private final GroupClient groupClient = new GroupClientImpl(JAVARUSH_URL);
 
     @BeforeAll
     static void mockClientSetup() {
-        client.expect(HttpMethod.GET, URL + GROUPS)
+        clientMock.expect(HttpMethod.GET, JAVARUSH_URL + GROUPS_PATH)
                 .thenReturn(List.of(new GroupInfo()));
-        client.expect(HttpMethod.GET, URL + GROUPS)
+        clientMock.expect(HttpMethod.GET, JAVARUSH_URL + GROUPS_PATH)
                 .queryString("limit", "2")
                 .thenReturn(List.of(new GroupInfo(), new GroupInfo()));
 
-        client.expect(HttpMethod.GET, String.format(GROUP_COUNT_PATH, URL + GROUPS))
+        clientMock.expect(HttpMethod.GET, String.format(GROUP_COUNT_PATH, JAVARUSH_URL + GROUPS_PATH))
                 .thenReturn(32);
-        client.expect(HttpMethod.GET, String.format(GROUP_COUNT_PATH, URL + GROUPS))
+        clientMock.expect(HttpMethod.GET, String.format(GROUP_COUNT_PATH, JAVARUSH_URL + GROUPS_PATH))
                 .queryString("type", "TECH")
                 .thenReturn(7);
 
@@ -47,16 +46,16 @@ class GroupClientImplTest {
         groupInfo.setId(16);
         groupInfo.setType(TECH);
         groupInfo.setKey("android");
-        client.expect(HttpMethod.GET, String.format(GROUP_BY_ID_PATH, URL + GROUPS, 16))
+        clientMock.expect(HttpMethod.GET, String.format(GROUP_BY_ID_PATH, JAVARUSH_URL + GROUPS_PATH, 16))
                 .thenReturn(groupInfo);
 
-        client.expect(HttpMethod.GET,  URL + POSTS)
+        clientMock.expect(HttpMethod.GET,  JAVARUSH_URL + POSTS_PATH)
                 .queryString("groupKid", "16")
                 .thenReturn(List.of());
 
         PostInfo post = new PostInfo();
         post.setId(1);
-        client.expect(HttpMethod.GET,  URL + POSTS)
+        clientMock.expect(HttpMethod.GET,  JAVARUSH_URL + POSTS_PATH)
                 .queryString("groupKid", "15")
                 .thenReturn(Collections.singletonList(post));
     }