+++ src/test/java/com/github/afanas10101111/jtb/command/HelpCommandTest.java
@@ -16,6 +16,6 @@ class HelpCommandTest extends AbstractCommandTest {
 
     @Override
     Command getCommand() {
-        return new HelpCommand(messageService);
+        return new HelpCommand(messageServiceMock);
     }
 }